Fix comments and do not re-do the delete if it has already been done
git-svn-id: https://svn.java.net/svn/glassfish~svn/trunk/main@60837 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5

+++ nucleus/common/common-util/src/main/java/com/sun/enterprise/util/io/FileUtils.java
@@ -1386,7 +1386,7 @@ public class FileUtils {
     }
     
     /**
-     * Retriable work for opening a FileOutputStream.
+     * Retriable work for deleting a file
      */
     private static class DeleteFileWork implements RetriableWork {
 
@@ -1399,6 +1399,8 @@ public class FileUtils {
 
         @Override
         public void run() {
+            if (complete) return;
+            
             if (deleteMe.delete()) complete = true;
         }
